Corrected compatibility problem between python 3 and 2

+++ pes/exercise_2/pre_processing.py
@@ -195,7 +195,7 @@ class CoulombMatrix():
                     'N': 7.0
                  }
 
-        self.n_atoms = len(self.rawX[0])/4
+        self.n_atoms = int(len(self.rawX[0])/4)
         self.n_samples = len(self.rawX)
